% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm_weights_oo.R
\docType{data}
\name{abm_weights_oo}
\alias{abm_weights_oo}
\title{abm_model_weights}
\format{
A data frame with 75200 rows and 5 variables:
\describe{
  \item{\code{ID}}{double agent ID}
  \item{\code{w_q14}}{double financial weight}
  \item{\code{w_q45}}{double social weight}
  \item{\code{w_theta}}{double weight}
  \item{\code{calibration}}{double calibration run (1-100)}
}
}
\usage{
abm_weights_oo
}
\description{
model weights in 100 micro-calibration runs
}
\details{
generated by gen_model_ensemble()
}
\keyword{datasets}
