% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finance.R
\name{amort}
\alias{amort}
\title{amort}
\usage{
amort(r, term)
}
\arguments{
\item{r}{interest rate}

\item{term}{term}
}
\value{
annual amortisation payment for unit principal
}
\description{
amort
}
\examples{
amort(0.07,20)
}
