% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finance.R
\name{annualised_system_cost}
\alias{annualised_system_cost}
\title{annualised_system_cost}
\usage{
annualised_system_cost(S, B, params)
}
\arguments{
\item{S}{total installed solar capacity}

\item{B}{storage capacity}

\item{params}{parameters created by scenario_params(sD,yeartime)}
}
\value{
annual cost in euros
}
\description{
this includes SEAI grant but excludes any resilience premium
}
