% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costs.R
\name{bess_cost_fun}
\alias{bess_cost_fun}
\title{bess_cost_fun}
\usage{
bess_cost_fun(sD, yeartime)
}
\arguments{
\item{sD}{scenario parameters dataframe}

\item{yeartime}{decimal time}
}
\value{
scalar cost in euros/kWh
}
\description{
solar battery cost in euros/kWh
}
