% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costs.R
\name{bess_install_cost_fun}
\alias{bess_install_cost_fun}
\title{bess_install_cost_fun}
\usage{
bess_install_cost_fun(sD, yeartime)
}
\arguments{
\item{sD}{scenario dataframe}

\item{yeartime}{decimal time}
}
\value{
scalar cost in euros
}
\description{
bess_install_cost_fun
}
