% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical.R
\name{bess_round}
\alias{bess_round}
\title{bess_round}
\usage{
bess_round(x, bess_increment = 2.5)
}
\arguments{
\item{x}{optimum storage}

\item{bess_increment}{minimum bess increment}
}
\description{
round to nearest multiple of bess_increment
}
