% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bill_values.R
\docType{data}
\name{bill_values}
\alias{bill_values}
\title{bill_values}
\format{
A data frame with 13 rows and 2 variables:
\describe{
  \item{\code{response_code}}{double response code}
  \item{\code{bill}}{character euros}
}
}
\usage{
bill_values
}
\description{
bi-monthly electricity bill conversion table
}
\details{
2024 survey
}
\keyword{datasets}
