% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{calABM}
\alias{calABM}
\title{calABM}
\usage{
calABM(sD, Nrun = 4, n_unused_cores = 2, use_parallel = T, beta, lambda, p, nu)
}
\arguments{
\item{sD}{base scenario (historical)}

\item{Nrun}{number of runs}

\item{n_unused_cores}{unsued cores default 2}

\item{use_parallel}{TRUE or FALSE}

\item{beta}{financial utility scale (drawn from financial_utility_scale)}

\item{lambda}{bias correction parameter}

\item{p}{rate parameter default 0.085}

\item{nu}{usable roof fraction for solar}
}
\description{
macro-calibration of pvbess
}
\details{
return uptake (number of households and installed capacity) in june 2023 and 2024 for comparison with ISEA reports
}
