% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finance.R
\name{clearness_index_fun}
\alias{clearness_index_fun}
\title{clearness_index_fun}
\usage{
clearness_index_fun(day, K_max, K_min, phase_K = 0, exponent = 4)
}
\arguments{
\item{day}{day of year 1..366}

\item{K_max}{max value of clearness index 0 < K_max < 1}

\item{K_min}{K_min < K_max}

\item{phase_K}{phase 0 if K_max occurs in mid summer, pi/2 if K_max occurs in mid winter}

\item{exponent}{flatness of clearness index in clear months, >= 2. For Dublin airport exponent is about 4.}
}
\value{
real number
}
\description{
describes the reduction in solar potential to the atmosphere and cloudiness as a function of day of year
}
\examples{
clearness_index_fun(30,0.45,0.3)
}
