% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finance.R
\name{day_length_fun}
\alias{day_length_fun}
\title{day_length_fun}
\usage{
day_length_fun(day, latitude)
}
\arguments{
\item{day}{day of year (1 .. 365)}

\item{latitude}{latitude angle in degrees}
}
\value{
hours
}
\description{
daylight hours as a function of day of year and latitude
}
\examples{
day_length_fun(1,53)
}
