% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finance.R
\name{daylight_usage_fun}
\alias{daylight_usage_fun}
\title{daylight_usage_fun}
\usage{
daylight_usage_fun(day, rho_solstice)
}
\arguments{
\item{day}{day of year}

\item{rho_solstice}{daylight usage on solstice (12 hours)}
}
\value{
a number between 0 and 1
}
\description{
daylight_usage_fun
}
\examples{
daylight_usage_fun(80,0.25)
}
