% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finance.R
\name{demand_fun}
\alias{demand_fun}
\title{demand_fun}
\usage{
demand_fun(day, D_max, D_min, lag_D = 30)
}
\arguments{
\item{day}{day of year}

\item{D_max}{maxiumum demand in kWh}

\item{D_min}{minimum demand in kWh}

\item{lag_D}{lag (in days) default 30}
}
\value{
daily kWh demand
}
\description{
A simple sinusoidal model of mean daily demand
}
\examples{
demand_fun(30,14,11,30 )
}
