% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_utils_oo.R
\docType{data}
\name{empirical_utils_oo}
\alias{empirical_utils_oo}
\title{empirical_utils_oo}
\format{
A data frame with 1800 rows and 4 variables:
\describe{
  \item{\code{question_code}}{character COLUMN_DESCRIPTION}
  \item{\code{response_code}}{double COLUMN_DESCRIPTION}
  \item{\code{du_average}}{double mean utility}
  \item{\code{calibration}}{calibration run}
}
}
\usage{
empirical_utils_oo
}
\description{
empirical partial utilities for owner-occupiers. Initial bias-correction using epsilon=0.7.
}
\details{
generated by get_model_ensemble()
}
\keyword{datasets}
