% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finance.R
\name{energy_flows_fast}
\alias{energy_flows_fast}
\title{energy_flows_fast}
\usage{
energy_flows_fast(
  S_1,
  S_2,
  aspect,
  shading_factor_1 = 1,
  shading_factor_2 = 1,
  B,
  D_max = 18,
  D_min = 11,
  params
)
}
\arguments{
\item{S_1}{installed solar capacity primary roof}

\item{S_2}{installed solar capacity secondary roof}

\item{aspect}{house orientation}

\item{shading_factor_1}{shading factor primary roof section}

\item{shading_factor_2}{shading factor secondary roof section}

\item{B}{installed battery capacity}

\item{D_max}{max demand}

\item{D_min}{min demand}

\item{params}{parameters object created by scenario_params()}
}
\value{
a tibble (converted from a data table)
}
\description{
energy_flows_fast
}
