% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costs.R
\name{evening_tariff_fun}
\alias{evening_tariff_fun}
\title{evening_tariff_fun}
\usage{
evening_tariff_fun(sD, yeartime)
}
\arguments{
\item{sD}{scenario dataframe}

\item{yeartime}{decimal time}
}
\value{
price per kWh in euros
}
\description{
actual (currenly to mid 2023) and projected path of electricity prices. Data from seai_elec. For inflation expectations see electricity_price_inflation_fun.
}
