% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/financial_utility_scale.R
\docType{data}
\name{financial_utility_scale}
\alias{financial_utility_scale}
\title{financial utility scale}
\format{
A data frame with 1000 rows and 2 variables:
\describe{
  \item{\code{i}}{integer index from repeated runs of initialise_agents}
  \item{\code{scale}}{double beta. scale}
}
}
\usage{
financial_utility_scale
}
\description{
scale calculate by comparing 2024 pvbess cost savings with financial utility inferred from survey
}
\details{
DETAILS
}
\keyword{datasets}
