% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costs.R
\name{fit_fun}
\alias{fit_fun}
\title{fit_fun}
\usage{
fit_fun(sD, yeartime)
}
\arguments{
\item{sD}{scenario dataframe}

\item{yeartime}{decimal time}
}
\value{
euro/kWh
}
\description{
clean export guarantee price euro/kWh to 2022-2050
}
