% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costs.R
\name{fit_inflation_fun}
\alias{fit_inflation_fun}
\title{FiT_inflation_fun}
\usage{
fit_inflation_fun(sD, yeartime)
}
\arguments{
\item{sD}{scenario dataframe}

\item{yeartime}{decimal time}
}
\value{
fit inflation expectation in decimal units
}
\description{
feed-in tariff (ceg) price inflation expectations in decimal units. This is used in NPV calculations.
}
