% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical.R
\name{gen_microcalibration_ensemble}
\alias{gen_microcalibration_ensemble}
\title{gen_microcalibration_ensemble}
\usage{
gen_microcalibration_ensemble(N, regularisation = 1, epsilon = 0.7)
}
\arguments{
\item{N}{number of independent calibration runs}

\item{regularisation}{parameter controlling presence of negative weights 1= full, 0= none}

\item{epsilon}{initial hypothetical bias correction,1= no bias correction, default = 0.7}
}
\value{
a two-component list of the empirical utilities and weights
}
\description{
generates a calibrated abm N-member ensemble (model empirical utils and weights) based on functions defined
in pvbesscalibrater. The utilities and weights are indexed according to the calibration run.
}
