% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical.R
\name{gen_roofsection_solar_area}
\alias{gen_roofsection_solar_area}
\title{gen_roofsection_solar_area}
\usage{
gen_roofsection_solar_area(
  region = 1,
  house_type = 1,
  stories = 1,
  bedrooms = 2,
  house_period = 6,
  roof_floor_ratio = 2/sqrt(3),
  usable_roof_fraction = 0.7
)
}
\arguments{
\item{region}{region 1=Dublin 2= rest of Leinster 3= Munster 4= Ulster/Connaught}

\item{house_type}{1=flat 2= terraced 3= semi 4= detached 5=other}

\item{stories}{number of stories}

\item{bedrooms}{number of bedrooms}

\item{house_period}{construction year integer}

\item{roof_floor_ratio}{2/sqrt(3) for 30 degree pitch}

\item{usable_roof_fraction}{fraction of rooftop usable for solar PV default 0.7}
}
\value{
potential half-roof area for solar pv in m2
}
\description{
generate a half rooftop solar area potential stochastically (log normal distribution). This is calibrated to BER data up to 2023 separately for
bungalows and two-storey dwellings. Bungalows have a higher solar potential because of their relatively larger footprint. This function uses
a detailed model based on house location (region), housing_type (q1) and number of storeys (1 or 2).
}
