% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costs.R
\name{get_demand_params}
\alias{get_demand_params}
\title{get_demand_params}
\usage{
get_demand_params(highest_kwh, lowest_kwh, lag_D = 30)
}
\arguments{
\item{highest_kwh}{highest kWh bi-monthly usage inferred from 2023 bills}

\item{lowest_kwh}{lowest kWh bi-monthly usage inferred from 2023 bills}

\item{lag_D}{demand seasonal lag default 30 days}
}
\description{
utility function used by
}
