% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{get_financial_utility_scale}
\alias{get_financial_utility_scale}
\title{get_financial_utility_scale}
\usage{
get_financial_utility_scale(agents_in, cal_run)
}
\arguments{
\item{agents_in}{agent chacteristics e.g. output by initialise_agents()}

\item{cal_run}{calibration run in 1 to 100}
}
\value{
a scalar
}
\description{
get_financial_utility_scale
}
