% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homophily.R
\docType{data}
\name{homophily}
\alias{homophily}
\title{homophily}
\format{
A data frame with 7 rows and 2 variables:
\describe{
  \item{\code{variable}}{character COLUMN_DESCRIPTION}
  \item{\code{weights}}{double COLUMN_DESCRIPTION}
}
}
\usage{
homophily
}
\description{
weights for link probability in make_artificial_society
}
\details{
DETAILS
}
\keyword{datasets}
