% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finance.R
\name{lcoe_pv}
\alias{lcoe_pv}
\title{lcoe_pv}
\usage{
lcoe_pv(mean_daily_potential, mu)
}
\arguments{
\item{mean_daily_potential}{annual average solar potentials (hours)}

\item{mu}{annualised capital cost of solar per kW installed (may include grant)}
}
\value{
euros/kWh
}
\description{
levelised cost of pvbess based on annualised system cost calculated using annualised_system_cost
}
\examples{
lcoe_pv(2.5,40)
}
