% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finance.R
\name{lcos_bess}
\alias{lcos_bess}
\title{lcos_bess}
\usage{
lcos_bess(B, lambda)
}
\arguments{
\item{B}{bess capacity (kWh)}

\item{lambda}{annualised cost (B kWh)}
}
\description{
levelised cost of storage
}
\details{
This is function of B
}
