% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costs.R
\name{night_tariff_fun}
\alias{night_tariff_fun}
\title{night_tariff_fun}
\usage{
night_tariff_fun(sD, yeartime)
}
\arguments{
\item{sD}{scenario dataframe}

\item{yeartime}{decimal time}
}
\value{
price per kWh in euros
}
\description{
Night tariffs. Historical rates are assumed to be 45% of the day rate
}
