% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pv_qanda.R
\docType{data}
\name{pv_qanda}
\alias{pv_qanda}
\title{pv_qanda}
\format{
A data frame with 256 rows and 4 variables:
\describe{
  \item{\code{question}}{character survey question}
  \item{\code{question_code}}{character question code}
  \item{\code{response}}{character reponse}
  \item{\code{response_code}}{double response code}
}
}
\usage{
pv_qanda
}
\description{
a table of response options for each pv_survey question
}
\details{
2024 Li et al
}
\keyword{datasets}
