% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pv_questions.R
\docType{data}
\name{pv_questions}
\alias{pv_questions}
\title{pv_questions}
\format{
A data frame with 43 rows and 2 variables:
\describe{
  \item{\code{question_code}}{character survey code}
  \item{\code{question}}{character survey question}
}
}
\usage{
pv_questions
}
\description{
survey features used in pvbessmicrosimr.
}
\details{
2024 Li et al
}
\keyword{datasets}
