% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pv_retrofit_uptake.R
\docType{data}
\name{pv_retrofit_uptake}
\alias{pv_retrofit_uptake}
\title{pv_retrofit_uptake}
\format{
A data frame with 15 rows and 2 variables:
\describe{
  \item{\code{date}}{double date}
  \item{\code{total}}{double cumulative total number of installatios}
}
}
\usage{
pv_retrofit_uptake
}
\description{
multi-source estimate of residential
}
\details{
inferred from BER dataset, 2022 Census, ISEA scale of solar reports for 2023 and 2024
}
\keyword{datasets}
