% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pv_society_oo.R
\docType{data}
\name{pv_society_oo}
\alias{pv_society_oo}
\title{pv_society_oo}
\format{
A data frame with 752 rows and 8 variables:
\describe{
  \item{\code{ID}}{integer COLUMN_DESCRIPTION}
  \item{\code{age}}{character COLUMN_DESCRIPTION}
  \item{\code{gender}}{character COLUMN_DESCRIPTION}
  \item{\code{region}}{character COLUMN_DESCRIPTION}
  \item{\code{class}}{character COLUMN_DESCRIPTION}
  \item{\code{education}}{character COLUMN_DESCRIPTION}
  \item{\code{area_type}}{character COLUMN_DESCRIPTION}
  \item{\code{degree}}{double COLUMN_DESCRIPTION}
}
}
\usage{
pv_society_oo
}
\description{
node attributes of artificial society
}
\details{
DETAILS
}
\keyword{datasets}
