% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pv_survey.R
\docType{data}
\name{pv_survey}
\alias{pv_survey}
\title{DATASET_TITLE}
\format{
A data frame with 726 rows and 43 variables:
\describe{
  \item{\code{serial}}{double COLUMN_DESCRIPTION}
  \item{\code{qa}}{double COLUMN_DESCRIPTION}
  \item{\code{qb}}{double COLUMN_DESCRIPTION}
  \item{\code{qc2}}{double COLUMN_DESCRIPTION}
  \item{\code{qd}}{double COLUMN_DESCRIPTION}
  \item{\code{qe}}{double COLUMN_DESCRIPTION}
  \item{\code{qf}}{double COLUMN_DESCRIPTION}
  \item{\code{q1}}{double COLUMN_DESCRIPTION}
  \item{\code{q2}}{double COLUMN_DESCRIPTION}
  \item{\code{q3}}{double COLUMN_DESCRIPTION}
  \item{\code{q4}}{double COLUMN_DESCRIPTION}
  \item{\code{q5}}{double COLUMN_DESCRIPTION}
  \item{\code{q6}}{double COLUMN_DESCRIPTION}
  \item{\code{q10}}{double COLUMN_DESCRIPTION}
  \item{\code{q14}}{double COLUMN_DESCRIPTION}
  \item{\code{q15}}{double COLUMN_DESCRIPTION}
  \item{\code{q261}}{double COLUMN_DESCRIPTION}
  \item{\code{q262}}{double COLUMN_DESCRIPTION}
  \item{\code{q263}}{double COLUMN_DESCRIPTION}
  \item{\code{q264}}{double COLUMN_DESCRIPTION}
  \item{\code{q265}}{double COLUMN_DESCRIPTION}
  \item{\code{q29}}{double COLUMN_DESCRIPTION}
  \item{\code{q30_1}}{double COLUMN_DESCRIPTION}
  \item{\code{q30_2}}{double COLUMN_DESCRIPTION}
  \item{\code{q30_3}}{double COLUMN_DESCRIPTION}
  \item{\code{q30_4}}{double COLUMN_DESCRIPTION}
  \item{\code{q30_5}}{double COLUMN_DESCRIPTION}
  \item{\code{q30_6}}{double COLUMN_DESCRIPTION}
  \item{\code{q36}}{double COLUMN_DESCRIPTION}
  \item{\code{q37}}{double COLUMN_DESCRIPTION}
  \item{\code{q38}}{double COLUMN_DESCRIPTION}
  \item{\code{q39}}{double COLUMN_DESCRIPTION}
  \item{\code{q40}}{double COLUMN_DESCRIPTION}
  \item{\code{q41}}{double COLUMN_DESCRIPTION}
  \item{\code{q42}}{double COLUMN_DESCRIPTION}
  \item{\code{q43}}{double COLUMN_DESCRIPTION}
  \item{\code{q44}}{double COLUMN_DESCRIPTION}
  \item{\code{q45}}{double COLUMN_DESCRIPTION}
  \item{\code{q46_5}}{double COLUMN_DESCRIPTION}
  \item{\code{qg}}{double COLUMN_DESCRIPTION}
  \item{\code{qh}}{double COLUMN_DESCRIPTION}
  \item{\code{qi}}{double COLUMN_DESCRIPTION}
  \item{\code{qj}}{double COLUMN_DESCRIPTION}
}
}
\usage{
pv_survey
}
\description{
DATASET_DESCRIPTION
}
\details{
DETAILS
}
\keyword{datasets}
