% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finance.R
\name{pvbess_cost_fun}
\alias{pvbess_cost_fun}
\title{pvbess_cost_fun}
\usage{
pvbess_cost_fun(
  S_1,
  S_2,
  aspect,
  shading_factor_1,
  shading_factor_2,
  B,
  D_max,
  D_min,
  params,
  tariff_plan = "night_saver"
)
}
\arguments{
\item{S_1}{installed PV capacity primary roof section}

\item{S_2}{installed PV capacity secondary roof section}

\item{aspect}{orientation of building}

\item{shading_factor_1}{primary shading factor}

\item{shading_factor_2}{secondary shading factor}

\item{B}{storage capacity}

\item{D_max}{max demand (winter)}

\item{D_min}{min demand (summer)}

\item{params}{parameter object created by scenario_params(sD, yeartime)}

\item{tariff_plan}{if "24hr" the night rate is set equal to the day  rate}
}
\value{
an annualised cost in euros
}
\description{
annualised capital and operational cost of a household energy system including grid, pvbess and finance costs
}
\examples{
pvbess_cost_fun(5,5,"SW-NE",1,1,5,18,11,scenario_params(sD,2025),"night_saver")
}
