% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finance.R
\name{pvbess_cost_fun_simple}
\alias{pvbess_cost_fun_simple}
\title{pvbess_cost_fun_simple}
\usage{
pvbess_cost_fun_simple(S, B, D_max, D_min, params, tariff_plan = "night_saver")
}
\arguments{
\item{S}{PV capacity}

\item{B}{storage capacity}

\item{D_max}{max demand (winter)}

\item{D_min}{min demand (summer)}

\item{params}{parameter object created by scenario_params(sD, yeartime)}

\item{tariff_plan}{if "24hr" the night rate is set equal to the day  rate}
}
\description{
annualised cost of residential electricity system include grid, pvbess and finance costs
}
