% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finance.R
\name{pvbess_imports_fun}
\alias{pvbess_imports_fun}
\title{pvbess_imports_fun}
\usage{
pvbess_imports_fun(
  S_1,
  S_2,
  aspect,
  shading_factor_1,
  shading_factor_2,
  B,
  D_max,
  D_min,
  params,
  tariff_plan = "night_saver"
)
}
\arguments{
\item{S_1}{installed PV capacity primary roof section}

\item{S_2}{installed PV capacity secondary roof section}

\item{aspect}{orientation of building}

\item{shading_factor_1}{primary shading factor}

\item{shading_factor_2}{secondary shading factor}

\item{B}{storage capacity}

\item{D_max}{max demand (winter)}

\item{D_min}{min demand (summer)}

\item{params}{parameter object created by scenario_params(sD, yeartime)}

\item{tariff_plan}{if "24hr" the night rate is set equal to the day  rate}
}
\value{
scalar (kWh)
}
\description{
annual imports from grid in kWh
}
