% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finance.R
\name{pvbess_optim}
\alias{pvbess_optim}
\title{pvbess_optim}
\usage{
pvbess_optim(D_max, D_min, S_constraint, params, tariff_plan = "night_saver")
}
\arguments{
\item{D_max}{max daily usage}

\item{D_min}{min daily usage}

\item{S_constraint}{max rooftop capacity available}

\item{params}{parameter object created by scenario_params_df(sD, yeartime)}

\item{tariff_plan}{default "night_saver". Alternative is "24hr".}
}
\value{
a tibble with parameters and B_optimal, S_optimal
}
\description{
find the optimised PVBESS system as a function of household energy demand parameters, tariffs and technology costs
}
