% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/finance.R
\name{pvbess_optim_complex}
\alias{pvbess_optim_complex}
\title{pvbess_optim_complex}
\usage{
pvbess_optim_complex(
  aspect,
  solar_constraint_1,
  solar_constraint_2,
  shading_factor_1,
  shading_factor_2,
  D_max,
  D_min,
  params,
  tariff_plan = "night_saver"
)
}
\arguments{
\item{aspect}{house orientation}

\item{solar_constraint_1}{primary roof capacity constraint}

\item{solar_constraint_2}{secondary roof capacity constraint}

\item{shading_factor_1}{primary roof shading in 0 to 1}

\item{shading_factor_2}{secondary roof shading}

\item{D_max}{winter demand max}

\item{D_min}{summer demand min}

\item{params}{scenario parameters}

\item{tariff_plan}{tariff plan "night-saver" or "24hr"}
}
\value{
tibble with optimium system, annualised cost savings and other parameters
}
\description{
pvbess_optim_complex
}
