% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvbessmicrosimr-package.R
\docType{package}
\name{pvbessmicrosimr-package}
\alias{pvbessmicrosimr}
\alias{pvbessmicrosimr-package}
\title{pvbessmicrosimr: This Package Simulates The Uptake Of Solar PV And Battery Storage by Households}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\author{
\strong{Maintainer}: Joseph Wheatley \email{joe.wheatley@ucd.ie}

}
\keyword{internal}
