% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costs.R
\name{resilience_premium_fun}
\alias{resilience_premium_fun}
\title{resilience_premium_fun}
\usage{
resilience_premium_fun(sD, yeartime)
}
\arguments{
\item{sD}{scenario}

\item{yeartime}{year fraction}
}
\description{
annual resilience premium (willingness-to-pay) for 1 day of battery storage
}
\examples{
resilience_premium_fun(sD,2024)
}
