% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{runABM}
\alias{runABM}
\title{runABM}
\usage{
runABM(
  sD,
  Nrun = 1,
  simulation_end = 2030,
  resample_society = F,
  n_unused_cores = 2,
  use_parallel = T,
  ignore_social = F,
  quiet = TRUE
)
}
\arguments{
\item{sD}{scenario set-up dataframe, typically read with readxlxs(...,sheet=scenario)}

\item{Nrun}{integer, number runs}

\item{simulation_end}{the final year of simulation of early termination is required}

\item{resample_society}{if TRUE resample pv_society with replacement to capture additional variability}

\item{n_unused_cores}{number of cores left unused in parallel/foreach. Recommended values 2 or 1.}

\item{use_parallel}{if TRUE uses multiple cores. Use FALSE for diagnostic runs}

\item{ignore_social}{if TRUE ignore social network effects. Default is FALSE}

\item{quiet}{if TRUE reduce messaging}
}
\value{
a three component list - simulation output, scenario setup, meta-parameters
}
\description{
Runs pv system adoption simulation on artificial society of 752 agents.
Each run is performed on an independently generated social network with randomisation from initialise_agents() and
with a random micro_calibration run (index 1..100)
}
\details{
Bi-monthly timesteps.
}
