% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical.R
\name{scenario_params}
\alias{scenario_params}
\title{scenario_params}
\usage{
scenario_params(sD, yeartime)
}
\arguments{
\item{sD}{scenario parameters e.g. scenario_0}

\item{yeartime}{decimal time}
}
\value{
long form dataframe containing parameter names and values
}
\description{
builds the complete parameter set at yeartime from scenario sD
}
