% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sD.R
\docType{data}
\name{sD}
\alias{sD}
\title{Scenario Design}
\format{
A data frame with 92 rows and 6 variables:
\describe{
  \item{\code{category}}{character parameter category}
  \item{\code{variable}}{character parameter description}
  \item{\code{parameter}}{character parameter name}
  \item{\code{value}}{double parameter value}
  \item{\code{unit}}{character parameter unit}
  \item{\code{comment}}{character comment if any}
}
}
\usage{
sD
}
\description{
technical paramaters and costs, tariff and grant assumptions
}
\details{
creatively designed
}
\keyword{datasets}
