% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{seai_grant_cpp}
\alias{seai_grant_cpp}
\title{seai_grant_cpp}
\arguments{
\item{params}{fast scenario parameters}

\item{s}{solar capacity in kW}

\item{b}{battery capacity in kWh}
}
\value{
grant amount in euros
}
\description{
The fast version of seai_grant
}
