% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solar_potentials.R
\docType{data}
\name{solar_potentials}
\alias{solar_potentials}
\title{solar_potentials}
\format{
A data frame with 8 rows and 2 variables:
\describe{
  \item{\code{azimuth}}{character compass point}
  \item{\code{potential}}{double in hours}
}
}
\usage{
solar_potentials
}
\description{
mean annual solar potential by azimuth for ac PV systems for a 30 degree pitch roof
}
\details{
PVWATTs at Birr
}
\keyword{datasets}
