% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/technical.R
\name{solar_round_down}
\alias{solar_round_down}
\title{solar_round_down}
\usage{
solar_round_down(x, solar_increment = 0.25)
}
\arguments{
\item{x}{optiumum S}

\item{solar_increment}{default 0.25}
}
\value{
capacity in kWh
}
\description{
round solar installation capacity down to increment of panel capacity
}
