% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costs.R
\name{survey_bills_to_kwh}
\alias{survey_bills_to_kwh}
\title{survey_bills_to_kwh}
\usage{
survey_bills_to_kwh(pv_data_in, lag_D = 30)
}
\arguments{
\item{pv_data_in}{survey data e.g. pv_survey_oo}

\item{lag_D}{seasonal lag in demand, default 30 days.}
}
\description{
a function that converts highest and lower bi-monthly bills from survey to daily D_max and D_min assuming
a seasonal demand lag_D. D_max and D_min are used to estimate the financial return on pv bess investment.
Missing data are imputed by default. An issue that needs to be addressed is "level pay"
}
