% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abm.R
\name{update_agents}
\alias{update_agents}
\title{update_agents}
\usage{
update_agents(
  sD,
  yeartime,
  agents_in,
  social_network,
  ignore_social = F,
  beta. = 0.2532785,
  lambda. = 0.105,
  p. = 0.0085,
  rho. = 0.35,
  cal_run
)
}
\arguments{
\item{sD}{scenario dataframe}

\item{yeartime}{decimal time}

\item{agents_in}{input agent dataframe}

\item{social_network}{artifical social network}

\item{ignore_social}{option to ignore social effects. Default is FALSE.}

\item{beta.}{financial utility scale factor}

\item{lambda.}{bias (fixed in macro-calibration)}

\item{p.}{speed parameter (fixed in macro-calibration)}

\item{rho.}{rho_solstice}

\item{cal_run}{microcalibration run index between 1 and 100}
}
\value{
updated agent dataframe
}
\description{
micro-simulation time-step updater
}
\details{
The workhorse ABM function.Within a scenario, does a single month update of the agent characteristics. A random sample of agents evaluates their economic and social
utilities. If these exceed their individual threshold an optimal PV_BESS system is adopted.
}
